Reorganize the default policy builder module so that it's easier to understand
