+++ server/custom/policy-builders/default/index.js
@@ -1,7 +1,6 @@
 //module example for helping to build policy tables.
 const initGroups = require('./initiateFunctionalGroups.js');
 const addHmiLevels = require('./addHmiLevels.js');
-const editAppPolicy = require('./editAppPolicy.js');
 
 module.exports = function (log) {
     //exported functions. these are required to implement
@@ -12,16 +11,14 @@ module.exports = function (log) {
             //permissions.vehicleDataNames is an array of vehicle data components that can be permitted
 
             //first, create all the function group objects so that permissions can exist inside them
-            const defaultGroups = initGroups.createDefaultFunctionalGroups();
-            return defaultGroups;
+            return initGroups.createDefaultFunctionalGroups();
         },
         createGroupPermissions: function () {
             //the functional groups we defined in initialFunctionGroups came back with information including the id, so the
             //functional groups can now be referenced correctly when adding permissions
 
             //go through and evaluate all the functional groups and give each group certain permissions
-            const permissions = initGroups.generatePermissions();
-            return permissions;
+            return initGroups.generatePermissions();
         },
         modifyFunctionalGroupObject: function (funcGroupObj) {
             //add HMI LEVELs to the functional group object
@@ -33,10 +30,9 @@ module.exports = function (log) {
             //before finding the permitted ones. modify the appPolicy object
             return;
         },
-        //TODO: remove the modules not needed anymore. try and bundle all info into one module
         createAppPolicyObject: function (appPolicy, appObj) {
             //assign functional groups and final edits
-            return editAppPolicy(appPolicy, appObj);
+            return initGroups.editAppPolicy(appPolicy, appObj);
         }
     };
 }
